requirements.txt
git+https://github.com/django-nonrel/django@nonrel-1.5
django-mongodb-engine==0.6.0
djangorestframework==3.2.5
djangotoolbox==1.8.0
flufl.enum==4.1
mock==1.3.0
nltk==3.1
pymongo==2.8
pytest==2.8.7
scipy=0.17.0
scikit-learn=0.17.1
requests==2.9.1
wheel==0.24.0
django-cors-headers==1.1.0
